Refactor card components: Move CardBase, CardTitle, and NumberInsideCards to a new directory for better organization. Update cardData.vue and cardTest.vue to reflect new import paths and simplify card structure. Adjust styles in cardTest.vue for improved layout.
